move the materialize function to init utilities
