+++ src/lightning/fabric/utilities/init.py
@@ -12,14 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import itertools
-from typing import Any, Callable, Dict, Optional, Sequence
+from typing import Any, Callable, Dict, Optional, Sequence, Union
 
 import torch
+from torch.nn import Module, Parameter
+from torch.optim import Optimizer
 from torch.overrides import TorchFunctionMode
 from typing_extensions import override
 
 from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_1
 from lightning.fabric.utilities.types import _DEVICE
+from lightning.fabric.utilities.rank_zero import rank_zero_warn
 
 
 # From https://lernapparat.de/faster-model-init by Thomas Viehmann
@@ -56,7 +59,7 @@ class _EmptyInit(TorchFunctionMode):
         return func(*args, **kwargs)
 
 
-def _materialize(module: torch.nn.Module, device: _DEVICE) -> None:
+def _materialize(module: Module, device: _DEVICE) -> None:
     """Materialize a module."""
     if not _TORCH_GREATER_EQUAL_2_1:
         raise RuntimeError("recurse=False requires torch 2.1")
@@ -69,8 +72,45 @@ def _materialize(module: torch.nn.Module, device: _DEVICE) -> None:
     module.reset_parameters()
 
 
-def _materialize_meta_tensors(module: torch.nn.Module, device: _DEVICE) -> None:
+def _materialize_meta_tensors(module: Module, device: _DEVICE) -> None:
     """Materialize all tensors in a given module."""
     for module in module.modules():
-        if any(t.is_meta for t in itertools.chain(module.parameters(recurse=False), module.buffers(recurse=False))):
+        if _has_meta_device_parameters_or_buffers(module, recurse=False):
             _materialize(module, device)
+
+
+def _materialize_distributed_module(module: Module, device: torch.device) -> None:
+    # Reference: https://github.com/pytorch/torchtitan/blob/main/docs/fsdp.md#meta-device-initialization
+    # TODO: Introduce `Fabric.materialize(module)` to give user control when materialization should happen
+    # TODO: Make `torchmetrics.Metric` compatible with the `to_empty()` + `reset_parameters()` semantics
+    if not _has_meta_device_parameters_or_buffers(module):
+        return
+
+    module.to_empty(device=device)  # has to be called on the root module
+
+    uninitialized_modules = set()
+    for submodule in module.modules():
+        if all(False for _ in itertools.chain(submodule.parameters(recurse=False), submodule.buffers(recurse=False))):
+            # module has no parameters or buffers
+            continue
+        if callable(reset_method := getattr(submodule, "reset_parameters", None)):
+            reset_method()
+        else:
+            uninitialized_modules.add(type(submodule).__name__)
+
+    if uninitialized_modules:
+        rank_zero_warn(
+            "Parameter initialization incomplete. The following modules have parameters or buffers with uninitialized"
+            " memory because they don't define a `reset_parameters()` method for re-initialization:"
+            f" {', '.join(uninitialized_modules)}"
+        )
+
+
+def _has_meta_device_parameters_or_buffers(obj: Union[Module, Optimizer], recurse: bool = True) -> bool:
+    if isinstance(obj, Optimizer):
+        return any(
+            t.is_meta for param_group in obj.param_groups for t in param_group["params"] if isinstance(t, Parameter)
+        )
+    if isinstance(obj, Module):
+        return any(t.is_meta for t in itertools.chain(obj.parameters(recurse=recurse), obj.buffers(recurse=recurse)))
+    raise TypeError(f"Expected `torch.nn.Module` or `torch.optim.Optimizer`, got: {type(obj).__name__}")
